Conflicts:
	build.sbt
	src/main/resources/application.conf
	src/main/scala/com/example/actors/routes/ApiRouteActor.scala
	src/main/scala/com/example/controller/UserController.scala
	src/main/scala/com/example/repository/UserDAO.scala

+++ src/main/scala/com/example/domain/Fruit.scala
@@ -3,7 +3,7 @@ package com.example.domain
 import spray.httpx.SprayJsonSupport
 import spray.json
 
-case class Fruit (name : String, peso: Int, quantity: Long, image: String)
+case class Fruit (name : String, peso: Int, quantity: Int, image: String)
 
 object Fruit extends json.DefaultJsonProtocol with SprayJsonSupport {
   implicit val userLoginFormat = jsonFormat4(Fruit.apply)
